biint for amount

+++ ponder.schema.ts
@@ -4,7 +4,7 @@ export const buyevent = onchainTable("buyFromListings", (t) => ({
   id: t.text().primaryKey(),
   tokenAddress: t.text().notNull(),
   to: t.text().notNull(),
-  amount: t.integer().notNull(),
+  amount: t.bigint().notNull(),
 }));
